% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCohortProfile.R
\name{calcCohortProfile}
\alias{calcCohortProfile}
\title{Combines a given retirement distribution and individual asset price (or efficiency)
profile to give the price (or efficiency) profile of a cohort of assets.
Depending on the given combination method either the weighted average method
is used or the fixed profile method.}
\usage{
calcCohortProfile(
  survivalValues,
  config,
  profileConversionFunction = NULL,
  discountRate = NULL
)
}
\arguments{
\item{survivalValues}{a vector of real numbers in the interval [0,1]
representing the survival function of a retirement distribution}

\item{config}{the configuration with the combination method, discount rate
inflation rate and any parameters for the profile function}

\item{profileConversionFunction}{one of (priceToEfficiencyProfile, efficiencyToPriceProfile)}
}
\value{
A vector of monotonically decreasing real numbers in the interval [0,1]
representing the price (or efficiency) profile of an asset vintage evaluated
at regular intervals of one period
}
\description{
Combines a given retirement distribution and individual asset price (or efficiency)
profile to give the price (or efficiency) profile of a cohort of assets.
Depending on the given combination method either the weighted average method
is used or the fixed profile method.
}
\examples{
values <- seq(1, 0, by=-0.1)
config <- list(combinationMethod="1")
calcCohortProfile(values, "lin", config)
}
