% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNetStock.R
\name{calcNetStock}
\alias{calcNetStock}
\title{Calculate the net stock by correcting the given GFCF values for the
combined age-price and retirement distribution}
\usage{
calcNetStock(gfcf, survivalValues, config)
}
\arguments{
\item{gfcf}{A numeric vector}

\item{survivalValues}{A data.frame with the following fields: id, vintageId
and values (a vector of survival values)}

\item{config}{a configuration list}
}
\value{
A numeric vector
}
\description{
Calculate the net stock by correcting the given GFCF values for the
combined age-price and retirement distribution
}
