% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProductiveStock.R
\name{calcProductiveStock}
\alias{calcProductiveStock}
\title{Calculate the productive stock by correcting the given GFCF values for the
combined age-efficiency and retirement distribution. This function is almost
identical to \code{\link{CalcNetStock}}. The duplication is deliberate in case
there might be Productive Stock specific changes in the future.}
\usage{
calcProductiveStock(gfcf, survivalValues, config)
}
\arguments{
\item{gfcf}{A numeric vector}

\item{survivalValues}{A data.frame with the following fields: id, vintageId
and values (a vector of survival values)}

\item{config}{a configuration list}
}
\value{
A numeric vector
}
\description{
Calculate the productive stock by correcting the given GFCF values for the
combined age-efficiency and retirement distribution. This function is almost
identical to \code{\link{CalcNetStock}}. The duplication is deliberate in case
there might be Productive Stock specific changes in the future.
}
