% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStock.R
\name{calcStock}
\alias{calcStock}
\title{correct the given GFCF values for the combined age-efficiency and
retirement distribution.}
\usage{
calcStock(gfcf, survivalValues, config, profileConversionFunction = NULL)
}
\arguments{
\item{gfcf}{A numeric vector}

\item{survivalValues}{A data.frame with the following fields: id, vintageId
and values (a vector of survival values)}

\item{config}{a configuration list}

\item{profileConversionFunction}{one of (priceToEfficiencyProfile, efficiencyToPriceProfile)}
}
\value{
A numeric vector
}
\description{
correct the given GFCF values for the combined age-efficiency and
retirement distribution.
}
