% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{calcSurvivalValues}
\alias{calcSurvivalValues}
\title{Calculate the survival values for each vintage given
life length parameters and retirement distribution}
\usage{
calcSurvivalValues(lifeLengths, config)
}
\arguments{
\item{lifeLengths}{data.frame with life length mean (Average),
coefficient of variation (CoV), minimum (Min) and maximum (Max) columns}

\item{config}{configuration}
}
\value{
data.frame with the survival values as a list-column (values),
a vintage id (vintageId) and an id for each unique combination of the
parameters (id)
}
\description{
Calculate the survival values for each vintage given
life length parameters and retirement distribution
}
\examples{
ll <- data.frame(Min=2, Max=15, Average=7, CoV=3/7)
config <- list(retirementDist="pnorm", rightTruncate=FALSE,
offSet=0)
calcSurvivalValues(ll, config)
}
