% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCohortProfile.R
\name{cohortProfileTwo}
\alias{cohortProfileTwo}
\title{Combines a given retirement distribution and individual asset price (or efficiency)
profile to give the price (or efficiency) profile of a cohort of assets. This version
of the function uses the fixed profile method of combining the two inputs.}
\usage{
cohortProfileTwo(
  survivalValues,
  profileFunction,
  profileConversionFunction = NULL,
  discountRate = NULL,
  inflationRate = NULL,
  profileFunctionParam = 0
)
}
\arguments{
\item{survivalValues}{a vector of real numbers in the interval [0,1]
representing the survival function of a retirement distribution}

\item{profileFunction}{a functional form of a single asset price (or efficiency) profile}

\item{profileConversionFunction}{one of (priceToEfficiencyProfile, efficiencyToPriceProfile)}

\item{discountRate}{a real number between [0,1]}

\item{inflationRate}{a real number between [0,1]}

\item{profileFunctionParam}{any parameter to be passed on to the profile function}
}
\value{
A vector of monotonically decreasing real numbers in the interval [0,1]
representing the price (or efficiency) profile of an asset vintage evaluated
at regular intervals of one period
}
\description{
Combines a given retirement distribution and individual asset price (or efficiency)
profile to give the price (or efficiency) profile of a cohort of assets. This version
of the function uses the fixed profile method of combining the two inputs.
}
\examples{
cohortProfileTwo(seq(1, 0, -0.1), "lin")
}
