% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{const}
\alias{const}
\title{Constant profile function: asset retains full value (productive capacity)
until the point where it retires (given by T)}
\usage{
const(numPeriods, retirementAge, dummy = NULL)
}
\arguments{
\item{numPeriods}{number of periods after which to evaluate value (productive capacity)}

\item{retirementAge}{The number of periods after which the value (productive capacity)
will be zero}
}
\value{
1 if numPeriods < retirementAge and 0 otherwise
}
\description{
Constant profile function: asset retains full value (productive capacity)
until the point where it retires (given by T)
}
