% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.R
\name{convertProfileToMatrix}
\alias{convertProfileToMatrix}
\title{Convert the given data.frame containing vectors to a matrix where each vector
is shifted forwards by it's id}
\usage{
convertProfileToMatrix(profileValues)
}
\arguments{
\item{profileValues}{A data.frame with the following fields:
vintageId (unique id in ascending order) and values (vectors of equal length)}
}
\value{
a matrix
}
\description{
Convert the given data.frame containing vectors to a matrix where each vector
is shifted forwards by it's id
}
\examples{
df <- purrr::by_row(data.frame(vintageId=1:4), ~1:4, .to="values")
convertProfileToMatrix(df)
}
