% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{db}
\alias{db}
\title{Declining Balance profile function: similar to geometric decay but now rate of decay
represents the declining balance rate}
\usage{
db(numPeriods, retirementAge, rateOfDecay)
}
\arguments{
\item{numPeriods}{number of periods after which to evaluate value (productive capacity)}

\item{retirementAge}{The number of periods after which the value (productive capacity)
will be zero}

\item{rateOfDecay}{rate at which asset depreciates (deteriorates)}
}
\value{
real value between [1, 0]
}
\description{
Declining Balance profile function: similar to geometric decay but now rate of decay
represents the declining balance rate
}
