% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileConversion.R
\name{efficiencyToPriceProfile}
\alias{efficiencyToPriceProfile}
\title{Convert a efficiency profile to a price (depreciation) profile}
\usage{
efficiencyToPriceProfile(profile, discountRate, inflationRate)
}
\arguments{
\item{profile}{a vector with values from the efficiency profile function}

\item{discountRate}{a real number between [0,1]}

\item{inflationRate}{a real number between [0,1]}
}
\value{
a vector with values from the price profile function
}
\description{
Convert a efficiency profile to a price (depreciation) profile
}
\examples{
efficiencyProf <- seq(1, 0, -0.125)
efficiencyToPriceProfile(efficiencyProf[-1], 0.05, 0.02)
}
