% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{geom}
\alias{geom}
\title{Geometric profile function: pattern of depreciation (deterioration) is geometric}
\usage{
geom(numPeriods, retirementAge, rateOfDecay)
}
\arguments{
\item{numPeriods}{number of periods after which to evaluate value (productive capacity)}

\item{retirementAge}{The number of periods after which the value (productive capacity)
will be zero}

\item{rateOfDecay}{rate at which asset depreciates (deteriorates)}
}
\value{
real value between [1, 0]
}
\description{
Geometric profile function: pattern of depreciation (deterioration) is geometric
}
