% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{getSurvivalValues}
\alias{getSurvivalValues}
\title{Calculate the survival values given a combination of parameters}
\usage{
getSurvivalValues(
  minimum,
  maximum,
  average,
  CoV,
  dist,
  rightTruncate,
  offSet,
  outputLength
)
}
\arguments{
\item{maximum}{the maximum life length of an asset}

\item{average}{average life length of an asset}

\item{CoV}{coefficient of variation of the retirement distribution of an asset}

\item{dist}{the retirement distribution, one of (pnorm, plnorm, pgamma, pweibull)}

\item{rightTruncate}{logical, if FALSE the distibution is not right truncated}

\item{offSet}{a real number between [-1, 1] indicating the point at which
the survival function should be evaluated}

\item{outputLength}{integer specifying the length of the output vector,
mainly to ensure survival vectors of different maximum life length to have
same length}

\item{mininum}{the minimum life length of an asset}
}
\value{
a vector of survival values with values between [0,1]
}
\description{
Calculate the survival values given a combination of parameters
}
\examples{
getSurvivalValues(2, 15, 7, 3/7, "pnorm", FALSE, 0, 15)
}
