% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{hyp}
\alias{hyp}
\title{Hyperbolic profile function: pattern of depreciation (deterioration) is slow
at first and increasing with the number of periods}
\usage{
hyp(numPeriods, retirementAge, shape)
}
\arguments{
\item{numPeriods}{number of periods after which to evaluate value (productive capacity)}

\item{retirementAge}{The number of periods after which the value (productive capacity)
will be zero}

\item{shape}{value determining the shape of the profile, must be between [0,1]}
}
\value{
real value between [1, 0]
}
\description{
Hyperbolic profile function: pattern of depreciation (deterioration) is slow
at first and increasing with the number of periods
}
