% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{lin}
\alias{lin}
\title{(Delayed) Linear profile function: value (productive capacity) of asset declines
linearly after a given number of periods (delay)}
\usage{
lin(numPeriods, retirementAge, delay = 0)
}
\arguments{
\item{numPeriods}{number of periods after which to evaluate value (productive capacity)}

\item{retirementAge}{The number of periods after which the value (productive capacity)
will be zero}

\item{delay}{the number of periods in which the asset retains its value (productive capacity)}
}
\value{
1 if numPeriods < delay, value between [1,0] otherwise
}
\description{
(Delayed) Linear profile function: value (productive capacity) of asset declines
linearly after a given number of periods (delay)
}
