% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileFunctions.R
\name{phyperform}
\alias{phyperform}
\title{Distribution function for a hyperbolic curve}
\usage{
phyperform(x, maximum, shape)
}
\arguments{
\item{x}{real valued scalar}

\item{maximum}{the value at which the distribution is right-truncated}

\item{shape}{value determining the shape of the profile, must be between [0,1]}
}
\value{
real value between [0,1]
}
\description{
Distribution function for a hyperbolic curve
}
