% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.R
\name{pim}
\alias{pim}
\title{Apply the perpetual inventory method to a time series of GCFC}
\usage{
pim(gfcf, profile)
}
\arguments{
\item{gfcf}{A numeric vector}

\item{profile}{A data.frame with the following fields:
vintageId (unique id in ascending order) and values (vectors of equal length)}
}
\value{
a numeric vector
}
\description{
Apply the perpetual inventory method to a time series of GCFC
}
\examples{
gfcf <- c(10, 20, 30, 40)
df <- purrr::by_row(data.frame(vintageId=1:4), ~1:4, .to="values")
pim(gfcf, df)
}
