% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimConfig.R, R/pimConfig_test.R
\name{pimConfig}
\alias{pimConfig}
\title{Configuration parameters for the calculation of gross/productive/net stock
using the perpetual inventory method}
\usage{
pimConfig(
  profileType = c("age-efficiency", "age-price"),
  profileFunctionName = c("lin", "geom", "const", "db", "hyp"),
  retirementDistName = c("pnorm", "pweibull", "plnorm", "pgamma", "none"),
  rightTruncate = FALSE,
  combinationMethod = c("1", "2"),
  discountRate = NULL,
  inflationRate = NULL,
  offSet = 0,
  profileFunctionParam = 0,
  discountRates = NULL
)

pimConfig(
  profileType = c("age-efficiency", "age-price"),
  profileFunctionName = c("lin", "geom", "const", "db", "hyp"),
  retirementDistName = c("pnorm", "pweibull", "plnorm", "pgamma", "none"),
  rightTruncate = FALSE,
  combinationMethod = c("1", "2"),
  discountRate = NULL,
  inflationRate = NULL,
  offSet = 0,
  profileFunctionParam = 0,
  discountRates = NULL
)
}
\arguments{
\item{profileType}{the type of profile: age-efficiency or age-price}

\item{profileFunctionName}{the name of a functional form of a single asset
price (or efficiency) profile. Options are linear (lin), geometric (geom),
constant (const), declining balance (db) or hyperbolic (hp)}

\item{retirementDistName}{the function name of the retirement distribution.
Options are Normal (pnorm), Weibull (pweibull), Log-Normal (plnorm),
Gamma (pgamma), or None (none)}

\item{rightTruncate}{logical value indicating if the retirement distribution
should be right truncated (TRUE) or not (FALSE)}

\item{combinationMethod}{the method employed to combine the age-efficieny/age-price
profile with the retirement distribution. This can either be the weighted
average ("1") or fixed profile ("2")}

\item{discountRate}{a real number between [0,1]}

\item{inflationRate}{a real number between [0,1]}

\item{offSet}{a real number between [0,1] representing the period at which
gfcf occurs and at which survival probabilities are evaluated}

\item{profileFunctionParam}{any parameters to be passed on to the profile function}

\item{discountRates}{a time series of discount rates with the same length as the
time series of GFCF. When set this value will take precedent over the single discount rate.}
}
\value{
a list with all the input values

a list with all the input values
}
\description{
Configuration parameters for the calculation of gross/productive/net stock
using the perpetual inventory method

Configuration parameters for the calculation of gross/productive/net stock
using the perpetual inventory method
}
