% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileConversion.R
\name{priceToEfficiencyProfile}
\alias{priceToEfficiencyProfile}
\title{Convert a price (depreciation)profile to a efficiency profile}
\usage{
priceToEfficiencyProfile(profile, discountRate, inflationRate)
}
\arguments{
\item{profile}{a vector with values from the price profile function}

\item{discountRate}{a real number between [0,1]}

\item{inflationRate}{a real number between [0,1]}
}
\value{
a vector with values from the price profile function
}
\description{
Convert a price (depreciation)profile to a efficiency profile
}
\examples{
agePriceProfile <- c(1, 0.785041, 0.594306, 0.428507, 0.288378, 0.174675,
0.088173, 0.029674, 0)
priceToEfficiencyProfile(agePriceProfile, 0.05, 0.02)
}
