% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Apply the Perpetual Inventory Method to a timeseries of GFCF to obtain a
timeseries of gross/productive/net stock}
\usage{
run(gfcf, lifeLengths, config)
}
\arguments{
\item{gfcf}{A numeric vector}

\item{lifeLengths}{lifeLengths data.frame with life length mean (Average),
coefficient of variation (CoV), minimum (Min) and maximum (Max) columns}

\item{config}{a list of configuration parameters (use the pimConfig function)}
}
\value{
a data.frame containing the gross stock (GrossStock), net stock (NetStock) and
productive stock (ProductiveStock)
}
\description{
Apply the Perpetual Inventory Method to a timeseries of GFCF to obtain a
timeseries of gross/productive/net stock
}
