% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformGammaParameters}
\alias{transformGammaParameters}
\title{Returns the shape and rate parameters of a Gamma distribution based on the
given parameters}
\usage{
transformGammaParameters(distMean, distCOV)
}
\arguments{
\item{distMean}{the mean of a Gamma distribution}

\item{distCOV}{the coefficient of variation of a Gamma distribution}
}
\value{
a vector of length two with elements shape and rate
}
\description{
Returns the shape and rate parameters of a Gamma distribution based on the
given parameters
}
