% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformLogNormalParameters}
\alias{transformLogNormalParameters}
\title{Returns the mean and standard deviation of a Log Normal distribution based
on the given parameters}
\usage{
transformLogNormalParameters(distMean, distCOV)
}
\arguments{
\item{distMean}{the mean of a Log Normal distribution}

\item{distCOV}{the coefficient of variation of a Log Normal distribution}
}
\value{
a vector of length two with elements meanLog and sdLog
}
\description{
Returns the mean and standard deviation of a Log Normal distribution based
on the given parameters
}
