% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformNoneParameters}
\alias{transformNoneParameters}
\title{Returns the parameters required to produce a "none" distribution from punif
This is a uniform distribution where min = 1 and max = 1}
\usage{
transformNoneParameters()
}
\value{
a vector of length two with elements min = 1 and max = 1
}
\description{
Returns the parameters required to produce a "none" distribution from punif
This is a uniform distribution where min = 1 and max = 1
}
