% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformNormalParameters}
\alias{transformNormalParameters}
\title{Returns the mean and standard deviation of a Normal distribution based on
the given parameters}
\usage{
transformNormalParameters(distMean, distCOV)
}
\arguments{
\item{distMean}{the mean of a Normal distribution}

\item{distCOV}{the coefficient of variation of a Normal distribution}
}
\value{
a vector of length two with elements distMean and distSD
}
\description{
Returns the mean and standard deviation of a Normal distribution based on
the given parameters
}
