% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformWeibullParameters}
\alias{transformWeibullParameters}
\title{Returns the shape and scale parameters of a Weibull distribution based on
the given parameters}
\usage{
transformWeibullParameters(distMean, distCOV)
}
\arguments{
\item{distMean}{the mean of a Weibull distribution}

\item{distCOV}{the coefficient of variation of a Weibull distribution}
}
\value{
a vector of length two with elements shape and scale
}
\description{
Returns the shape and scale parameters of a Weibull distribution based on
the given parameters
}
