% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvivalValues.R
\name{transformWeibullParametersInverse}
\alias{transformWeibullParametersInverse}
\title{The inverse of transformWeibullParameter: for a given shape and scale this
gives the mean and coefficient of variation of the corresponding Weibull
distribution}
\usage{
transformWeibullParametersInverse(scale, shape)
}
\arguments{
\item{scale}{real value}

\item{shape}{real value}
}
\value{
a vector of length two with elements distMean and distCOV
}
\description{
The inverse of transformWeibullParameter: for a given shape and scale this
gives the mean and coefficient of variation of the corresponding Weibull
distribution
}
